backend/requirements.txt
asgiref==3.9.1
autopep8==2.3.2
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-dotenv==1.4.2
django-environ==0.12.0
djangorestframework==3.16.0
idna==3.10
pycodestyle==2.14.0
requests==2.32.4
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
